requirements.txt
httpx
requests
pandas
numpy
matplotlib
seaborn
pymongo
praw
nltk
openai
reportlab
python-dotenv
python-dateutil
QuantLib